Form formatting and use one object per file
