+++ code/Form/CartForm.php
@@ -4,7 +4,6 @@ namespace SwipeStripe\Core\Form;
 
 use SS_Log;
 use Exception;
-
 use SilverStripe\Forms\FieldList;
 use SilverStripe\View\Requirements;
 use SwipeStripe\Core\code\Customer\Cart;
@@ -16,7 +15,6 @@ use SilverStripe\ORM\DataObject;
 use SilverStripe\Dev\Debug;
 use SwipeStripe\Core\code\Form\CartForm_QuantityField;
 use SwipeStripe\Core\code\Order\Item;
-use SilverStripe\Forms\TextField;
 
 /**
  * Form to display the {@link Order} contents on the {@link CartPage}.
@@ -28,14 +26,13 @@ use SilverStripe\Forms\TextField;
  */
 class CartForm extends Form
 {
-    
     /**
      * The current {@link Order} (cart).
      *
      * @var Order
      */
     public $order;
-    
+
     /**
      * Construct the form, set the current order and the template to be used for rendering.
      *
@@ -72,7 +69,6 @@ class CartForm extends Form
      */
     public function setupFormErrors()
     {
-
         //Only run when fields exist
         if ($this->fields->exists()) {
             parent::setupFormErrors();
@@ -87,7 +83,7 @@ class CartForm extends Form
         if ($items) {
             foreach ($items as $item) {
                 $fields->push(CartForm_QuantityField::create(
-                    'Quantity['.$item->ID.']',
+                    'Quantity[' . $item->ID . ']',
                     $item->Quantity,
                     $item
                 ));
@@ -117,7 +113,7 @@ class CartForm extends Form
         $items = $this->order->Items();
         if ($items) {
             foreach ($items as $item) {
-                $validator->addRequiredField('Quantity['.$item->ID.']');
+                $validator->addRequiredField('Quantity[' . $item->ID . ']');
             }
         }
 
@@ -147,14 +143,13 @@ class CartForm extends Form
     public function goToCheckout(array $data, Form $form)
     {
         $this->saveCart($data, $form);
-        
+
         if ($checkoutPage = DataObject::get_one(CheckoutPage::class)) {
             $this->controller->redirect($checkoutPage->AbsoluteLink());
         } else {
             Debug::friendlyError(500);
         }
     }
-
 
     /**
      * Save the cart, update the order item quantities and the order total.
@@ -165,7 +160,7 @@ class CartForm extends Form
     private function saveCart(array $data, Form $form)
     {
         $currentOrder = Cart::get_current_order();
-        $quantities = (isset($data['Quantity'])) ?$data['Quantity'] :null;
+        $quantities = (isset($data['Quantity'])) ? $data['Quantity'] : null;
 
         if ($quantities) {
             foreach ($quantities as $itemID => $quantity) {
@@ -192,176 +187,5 @@ class CartForm extends Form
     public function Cart()
     {
         return $this->order;
-    }
-}
-
-/**
- * Quantity field for displaying each {@link Item} in an {@link Order} on the {@link CartPage}.
- */
-class CartForm_QuantityField extends TextField
-{
-
-    /**
-     * Template for rendering the field
-     *
-     * @var String
-     */
-    protected $template = CartForm_QuantityField::class;
-    
-    /**
-     * Current {@link Item} represented by this field.
-     *
-     *  @var Item
-     */
-    protected $item;
-    
-    /**
-     * Construct the field and set the current {@link Item} that this field represents.
-     *
-     * @param String $name
-     * @param String $title
-     * @param String $value
-     * @param Int $maxLength
-     * @param Form $form
-     * @param Item $item
-     */
-    public function __construct($name, $value = "", $item = null)
-    {
-        $this->item = $item;
-        parent::__construct($name, '', $value, null, null);
-    }
-    
-    /**
-     * Render the field with the appropriate template.
-     *
-     * @see FormField::FieldHolder()
-     */
-    public function FieldHolder($properties = array())
-    {
-        $obj = ($properties) ? $this->customise($properties) : $this;
-        return $this->renderWith($this->template);
-    }
-    
-    /**
-     * Retrieve the current {@link Item} this field represents. Used in the template.
-     *
-     * @return Item
-     */
-    public function Item()
-    {
-        return $this->item;
-    }
-    
-    /**
-     * Set the current {@link Item} this field represents
-     *
-     * @param Item $item
-     */
-    public function setItem(Item $item)
-    {
-        $this->item = $item;
-    }
-    
-    /**
-     * Validate this field, check that the current {@link Item} is in the current
-     * {@Link Order} and is valid for adding to the cart.
-     *
-     * @see FormField::validate()
-     * @return Boolean
-     */
-    public function validate($validator)
-    {
-        $valid = true;
-        $item = $this->Item();
-        $currentOrder = Cart::get_current_order();
-        $items = $currentOrder->Items();
-        $quantity = $this->Value();
-
-        $removingItem = false;
-        if ($quantity <= 0) {
-            $removingItem = true;
-        }
-
-        //Check that item exists and is in the current order
-        if (!$item || !$item->exists() || !$items->find('ID', $item->ID)) {
-            $errorMessage = _t('Form.ITEM_IS_NOT_IN_ORDER', 'This product is not in the Cart.');
-            if ($msg = $this->getCustomValidationMessage()) {
-                $errorMessage = $msg;
-            }
-            
-            $validator->validationError(
-                $this->getName(),
-                $errorMessage,
-                "error"
-            );
-            $valid = false;
-        } elseif ($item) {
-
-            //If removing item, cannot subtract past 0
-            if ($removingItem) {
-                if ($quantity < 0) {
-                    $errorMessage = _t('Form.ITEM_QUANTITY_LESS_ONE', 'The quantity must be at least 0');
-                    if ($msg = $this->getCustomValidationMessage()) {
-                        $errorMessage = $msg;
-                    }
-                    
-                    $validator->validationError(
-                        $this->getName(),
-                        $errorMessage,
-                        "error"
-                    );
-                    $valid = false;
-                }
-            } else {
-                //If quantity is invalid
-                if ($quantity == null || !is_numeric($quantity)) {
-                    $errorMessage = _t('Form.ITEM_QUANTITY_INCORRECT', 'The quantity must be a number');
-                    if ($msg = $this->getCustomValidationMessage()) {
-                        $errorMessage = $msg;
-                    }
-                    
-                    $validator->validationError(
-                        $this->getName(),
-                        $errorMessage,
-                        "error"
-                    );
-                    $valid = false;
-                } elseif ($quantity > 2147483647) {
-                    $errorMessage = _t('Form.ITEM_QUANTITY_INCORRECT', 'The quantity must be less than 2,147,483,647');
-                    if ($msg = $this->getCustomValidationMessage()) {
-                        $errorMessage = $msg;
-                    }
-                    
-                    $validator->validationError(
-                        $this->getName(),
-                        $errorMessage,
-                        "error"
-                    );
-                    $valid = false;
-                }
-
-                $validation = $item->validateForCart();
-                if (!$validation->valid()) {
-                    $errorMessage = $validation->message();
-                    if ($msg = $this->getCustomValidationMessage()) {
-                        $errorMessage = $msg;
-                    }
-                    
-                    $validator->validationError(
-                        $this->getName(),
-                        $errorMessage,
-                        "error"
-                    );
-                    $valid = false;
-                }
-            }
-        }
-        
-        return $valid;
-    }
-
-    public function Type()
-    {
-        return 'cartquantity';
     }
 }
