+++ code/Mails/NotificationEmail.php
@@ -3,8 +3,8 @@
 namespace SwipeStripe\Core\Mails;
 
 use SilverStripe\Security\Member;
-use SwipeStripe\Core\code\Order\Order;
-use SwipeStripe\Core\code\Admin\ShopConfig;
+use SwipeStripe\Core\Order\Order;
+use SwipeStripe\Core\Admin\ShopConfig;
 use SilverStripe\Control\Email\Email;
 use SilverStripe\Control\Director;
 
@@ -19,7 +19,6 @@ use SilverStripe\Control\Director;
  */
 class NotificationEmail extends ProcessedEmail
 {
-
     /**
      * Create the new notification email.
      *
@@ -40,12 +39,12 @@ class NotificationEmail extends ProcessedEmail
             $this->to = $siteConfig->NotificationTo;
         }
         if ($siteConfig->NotificationSubject) {
-            $this->subject = $siteConfig->NotificationSubject . ' - Order #'.$order->ID;
+            $this->subject = $siteConfig->NotificationSubject . ' - Order #' . $order->ID;
         }
         if ($siteConfig->NotificationBody) {
             $this->body = $siteConfig->NotificationBody;
         }
-        
+
         if ($customer->Email) {
             $this->from = $customer->Email;
         } elseif (Email::getAdminEmail()) {
@@ -53,28 +52,28 @@ class NotificationEmail extends ProcessedEmail
         } else {
             $this->from = 'no-reply@' . $_SERVER['HTTP_HOST'];
         }
-        
+
         $this->signature = '';
         $adminLink = Director::absoluteURL('/admin/shop/');
 
         //Get css for Email by reading css file and put css inline for emogrification
         $this->setTemplate('Order_NotificationEmail');
-        
-        if (file_exists(Director::getAbsFile($this->ThemeDir().'/css/ShopEmail.css'))) {
-            $css = file_get_contents(Director::getAbsFile($this->ThemeDir().'/css/ShopEmail.css'));
+
+        if (file_exists(Director::getAbsFile($this->ThemeDir() . '/css/ShopEmail.css'))) {
+            $css = file_get_contents(Director::getAbsFile($this->ThemeDir() . '/css/ShopEmail.css'));
         } else {
             $css = file_get_contents(Director::getAbsFile('swipestripe/css/ShopEmail.css'));
         }
 
         $this->populateTemplate(
-            array(
+            [
                 'Message' => $this->Body(),
                 'Order' => $order,
                 'Customer' => $customer,
                 'InlineCSS' => "<style>$css</style>",
                 'Signature' => $this->signature,
                 'AdminLink' => $adminLink
-            )
+            ]
         );
         parent::__construct($from, null, $subject, $body, $bounceHandlerURL, $cc, $bcc);
     }
